refactor: rework project architecture
