+++ src/components/Button/Button.props.ts
@@ -1,0 +1,6 @@
+interface ButtonProps {
+  onClick: () => void;
+  label: string;
+}
+
+export default ButtonProps;
