+++ src/components/Button/ErrorButton.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const ErrorTrigger: React.FC = () => {
+  const triggerError = () => {
+    throw new Error('Error! Something went wrong.');
+  };
+
+  return (
+    <div>
+      <button onClick={triggerError}>Trigger Error</button>
+    </div>
+  );
+};
+
+export default ErrorTrigger;
