+++ src/components/Search/Search.props.tsx
@@ -1,0 +1,5 @@
+interface SearchProps {
+  onSearch: (searchTerm: string) => void;
+  initialSearchTerm: string;
+}
+export default SearchProps;
