--- src/components/TopControls.module.css
@@ -1,19 +1,0 @@
-.controls {
-  max-width: 600px;
-  margin: 50px auto;
-  padding: 20px;
-  text-align: center;
-}
-.btns {
-  display: flex;
-  justify-content: center;
-  gap: 20px;
-}   
-input {
-  padding: 12px 24px;
-  border: 1px solid #141885;
-  border-radius: 8px;
-  font-size: 1rem;
-  text-align: center;
-  margin-bottom: 20px;
-}
