+++ src/components/TopControls/TopControls.module.css
@@ -1,0 +1,14 @@
+input {
+    padding: 12px 24px;
+    border: 1px solid #141885;
+    border-radius: 8px;
+    font-size: 1rem;
+    text-align: center;
+    margin-bottom: 20px;
+  }
+  .controls {
+    max-width: 600px;
+    margin: 50px auto;
+    padding: 20px;
+    text-align: center;
+  }
