+++ src/components/TopControls/TopControls.props.ts
@@ -1,0 +1,5 @@
+interface TopControlsProps {
+  onSearch: (searchTerm: string) => void;
+  searchTerm: string;
+}
+export default TopControlsProps;
