+++ src/pages/HomePage/HomePage.module.css
@@ -1,0 +1,9 @@
+.containerHomePage {
+    background-color: #e1bffe;
+    padding: 20px;
+    margin: 0 auto;
+    max-width: 1200px;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    text-align: center;
+  }
